% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanText.R
\name{cleanText}
\alias{cleanText}
\title{Basic text cleaning}
\usage{
cleanText(text)
}
\arguments{
\item{text}{A character vector to clean.}
}
\value{
A character vector with normalized whitespace.
}
\description{
Lowercases text and removes punctuation, digits, and extra whitespace.
}
