% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_occs.R
\name{convertOccs}
\alias{convertOccs}
\title{Convert occupation codes}
\usage{
convertOccs(occs, from, to)
}
\arguments{
\item{occs}{A vector of occupation codes to convert.}

\item{from}{The source classification label.}

\item{to}{The target classification label.}
}
\value{
A vector of converted occupation codes.
}
\description{
Converts occupation codes between classifications using shortest paths.
}
