% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelWinDate.R
\name{excelWinDate}
\alias{excelWinDate}
\title{Convert Excel serial dates}
\usage{
excelWinDate(x)
}
\arguments{
\item{x}{Numeric vector of Excel serial dates.}
}
\value{
A `Date` vector.
}
\description{
Converts Windows Excel serial dates to `Date`.
}
