% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorInteractions.R
\name{factorInteractions}
\alias{factorInteractions}
\title{Combine factor levels}
\usage{
factorInteractions(x, y)
}
\arguments{
\item{x, y}{Factor vectors of equal length.}
}
\value{
A factor with combined levels.
}
\description{
Creates an interaction factor from two factors, preserving labels.
}
