% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatStandard.R
\name{formatStandard}
\alias{formatStandard}
\title{Standardize place-name formatting}
\usage{
formatStandard(x)
}
\arguments{
\item{x}{A character vector of names.}
}
\value{
A character vector of standardized names.
}
\description{
Cleans and standardizes geographic names with consistent capitalization.
}
