% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountry.R
\name{getCountry}
\alias{getCountry}
\title{Infer country from location strings}
\usage{
getCountry(locations)
}
\arguments{
\item{locations}{A character vector of location strings.}
}
\value{
A character vector of country names.
}
\description{
Geocodes locations and returns country names.
}
