% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDensities.R
\name{getDensities}
\alias{getDensities}
\title{Fetch population densities for coordinates}
\usage{
getDensities(latitudes, longitudes, n = 200)
}
\arguments{
\item{latitudes}{Numeric vector of latitudes.}

\item{longitudes}{Numeric vector of longitudes.}

\item{n}{Chunk size for API requests.}
}
\value{
A numeric vector of population densities.
}
\description{
Queries the Data Science Toolkit API for population density by coordinate.
}
