% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSheets.R
\name{listSheets}
\alias{listSheets}
\title{Read all sheets from an Excel file}
\usage{
listSheets(file)
}
\arguments{
\item{file}{Path to an Excel file.}
}
\value{
A named list of data frames.
}
\description{
Returns a named list of data frames, one per sheet.
}
