% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSankey.R
\name{makeSankey}
\alias{makeSankey}
\title{Create a Google Charts Sankey HTML file}
\usage{
makeSankey(switch, origin.date, dest.date, file = "sankey.html")
}
\arguments{
\item{switch}{A data frame with origin, destination, and percent columns.}

\item{origin.date}{Label to append to origin nodes.}

\item{dest.date}{Label to append to destination nodes.}

\item{file}{Output HTML file path.}
}
\value{
`NULL`, invisibly.
}
\description{
Writes an HTML file for a Sankey diagram using Google Visualization.
}
