% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distToReferencePoint.R
\name{manyDistToLoc}
\alias{manyDistToLoc}
\title{Distance to a reference point}
\usage{
manyDistToLoc(points, reference.point)
}
\arguments{
\item{points}{A two-column matrix/data frame of longitude and latitude.}

\item{reference.point}{A length-2 vector of longitude and latitude.}
}
\value{
A numeric vector of distances in kilometers.
}
\description{
Computes great-circle distances from multiple points to one reference point.
}
