% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mii.w6.model}
\alias{mii.w6.model}
\title{Issue coding SVM model}
\format{
An `svm` model object.
}
\usage{
mii.w6.model
}
\description{
Support vector machine model used for issue coding.
}
\keyword{datasets}
