% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minByRow.R
\name{minByRow}
\alias{minByRow}
\title{Row-wise minimum}
\usage{
minByRow(..., na.rm)
}
\arguments{
\item{...}{Numeric vectors or a matrix-like object to bind by column.}

\item{na.rm}{Logical; remove missing values.}
}
\value{
Numeric vector of row minima.
}
\description{
Compute row-wise minima across vectors or columns.
}
