% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occ.crosswalks}
\alias{occ.crosswalks}
\title{Occupation crosswalks}
\format{
A named list of crosswalk matrices.
}
\usage{
occ.crosswalks
}
\description{
Crosswalk tables for mapping occupation codes.
}
\keyword{datasets}
