% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2p.R
\name{p2p}
\alias{p2p}
\title{Proportion to percentage}
\usage{
p2p(x, round = 1)
}
\arguments{
\item{x}{Numeric vector of proportions.}

\item{round}{Number of decimal places to round to.}
}
\value{
Numeric vector of percentages.
}
\description{
Converts proportions to percentages.
}
