% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcileStrings.R
\name{reconcileStrings}
\alias{reconcileStrings}
\title{Reconcile string vectors}
\usage{
reconcileStrings(char.vector1, char.vector2, quiet = FALSE)
}
\arguments{
\item{char.vector1}{A character vector to recode.}

\item{char.vector2}{A character vector of target values.}

\item{quiet}{Logical; suppress progress output.}
}
\value{
A character vector with reconciled values.
}
\description{
Attempts to reconcile values in one character vector to match another.
}
