% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceValues.R
\name{replaceValues}
\alias{replaceValues}
\title{Replace tokens in text}
\usage{
replaceValues(text, subs)
}
\arguments{
\item{text}{A character vector of text.}

\item{subs}{A two-column matrix or data frame of replacements.}
}
\value{
A character vector with substitutions applied.
}
\description{
Replaces words in a text vector using a substitution table.
}
