% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spellCorrect.R
\name{spellCorrect}
\alias{spellCorrect}
\title{Approximate spell correction}
\usage{
spellCorrect(text, sensitivity = 0.11)
}
\arguments{
\item{text}{A character vector of text.}

\item{sensitivity}{Numeric threshold for edit-distance ratio.}
}
\value{
A list with corrected `text` and a `subs` matrix of replacements.
}
\description{
Suggests replacements based on edit distance across a text corpus.
}
