% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bes.R
\name{theme_bes}
\alias{theme_bes}
\title{BES ggplot theme}
\usage{
theme_bes(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme object.
}
\description{
Provides a simple ggplot theme with a white background.
}
