% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_data_2010}
\alias{us_data_2010}
\title{US data from 2010}
\format{
A list with elements `sample` and `discrete.targets`.
}
\usage{
us_data_2010
}
\description{
List data containing a `sample` data frame and `discrete.targets`.
}
\keyword{datasets}
