% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSentenceList.R
\name{writeSentenceList}
\alias{writeSentenceList}
\title{Format a list as a sentence}
\usage{
writeSentenceList(x, oxcomma = TRUE)
}
\arguments{
\item{x}{A character vector.}

\item{oxcomma}{Logical; include an Oxford comma if `TRUE`.}
}
\value{
A single character string.
}
\description{
Joins a character vector with commas and an optional Oxford comma.
}
